Course/DataMining/spider/NSFC_spider/requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytesseract==0.2.5
requests==2.19.1
six==1.11.0
tesseract==0.1.3
urllib3==1.23
Werkzeug==0.14.1
